// frontend/app/components/pagesComponent/sales/Cards/BasketCard/Interfaces/ClientDetails.ts
export interface ClientDetails {
    client_id: string;
    first_name: string;
    second_name: string;
    email: string;
    client_created_at: string;
    address_id: string;
    miasto: string;
    ulica: string;
    nr_budynku: string;
    nr_mieszkania: string;
    kod: string;
    nazwa_firmy: string;
    address_created_at: string;
}